CARGO SHIP MANAGEMENT SYSTEM By Vinay Edula
The Cargo Ship Management System is a simple PHP/MySQLi project that helps a cargo company or businesses manage their customers parcel details. The system stores all the branches or the company that can be also use when setting a destination where can recipient pickup their packages or cargos. The system has tracking feature where can help to monitor the movement of the customer'scargo. The system has 2 types of user which is the Admin  and the CSMS users. The Admin can manage all the data in the system including managing the branches and branches staffs. The users can only track their cargos and manage their list of parcels. The cargo items has multiple statuses which are the "Item Accepted ", "Collected", "Shipped", "In-Transit","Arrived At Destination", "Out for Delivery", "Ready to Pickup", "Delivered", "Picked-up", and "Unsuccessfull Delivery Attempt". This statuses will help to determin the cargo movement. The system also generates a report between two dates and selected status. The items of the clients can be set into "Deliver" and "Pickup". The cargos that marked as deliver are the items to be deliver directly to the recipient while the pickup will be delivered to the branch of the company near to the recipient address. The system admin or user can store or adds a multiple items at the same time but these items will be stored in the database seperately because each package has a different reference number or different tracking number. For example, user 1 has 3 boxes of package to be  sent to same recipient, the system user can submit the cargo booking registration to the system at once but will be stored seperately so that the system will generate a different unique reference number in each item so that they can track easily each items.